filter out only attnets nodes
Run method in Crawler now receives a sending channel so that the Crawler
can send the nodes back to the caller.

In the discv5-crawl command, it now runs another goroutine which runs
the Crawler and that goroutine sends the nodes back to the main
goroutine using a channel.

The main goroutine then iterates through the channel to get the nodes
and filter only for the nodes which have "attnets" key in the ENR.

We also create a new package "attnets" which contains things related to
"attnets" key in ENR.