create test framework and test Crawler
We use "go test" to test packages in the module and encapsulate it
inside the Makefile, so you can test the whole module by running "make
test".

There is a testlog package which is a facility to test whether the
code logs the correct things or not.

We add the tests for Crawler for both New and Run methods. However,
since we want to inject "disc" property inside Crawler, we need to
create a run method which is then called by Run. run is different from
Run in that it doesn't include a set-up phase for "disc".

We add crawler.Config which allows the users to customize more things in
the New method.

We also need to have "discv5" private interface because otherwise we
wouldn't be able to inject a fake discover.UDPv5 in the unit tests.